requirements.txt
langchain-community==0.2.5
langchain~=0.2.6
langchain-openai==0.1.14
tiktoken~=0.7.0
fastapi~=0.110.0
uvicorn~=0.28.0
pymilvus==2.4.0
langchain-cohere==0.1.8
requests~=2.32.0
cohere==5.5.8
tenacity==8.2.3
ragas==0.1.7
datasets~=3.0.0
python-dotenv~=1.0.1